hickey + lowy: one environment edge, and two files that are functions of values
Layer 2, fragmentation: `OLAI_IDENTITY_*` is one config family, and it was
being read in two modules under a rule nobody enforces — a fifth variable
would have had to remember both. It is now read in `config.ts` alone, which
makes that file the package's one edge (P2: effects at the boundary) and
leaves `identity.ts` a parse over names it is handed and `picture.ts` a ladder
over a template it is handed. Neither touches `process.env`, so their tests
state a deployment rather than arranging one, and the environment has exactly
one test (`config.test.ts`).

Lowy: the volatility is "how the deployment in front of olai says who somebody
is, and what they look like" — one axis, one socket (`IdentityConfig`), no
provider named anywhere in the code.
