+++ packages/identity/src/config.ts
@@ -1,21 +1,47 @@
 /**
- * Everything the operator said about identity, in one value.
+ * THE ENVIRONMENT EDGE: everything the operator said about identity, read
+ * once, as one value.
  *
- * The header NAMES ({@link ./identity.ts}) and the avatar template
- * ({@link ./picture.ts}) are one config family — `OLAI_IDENTITY_*`, read
- * once at the composition root and handed down — but they are not the same
- * KIND of thing, which is why they are two files and this is the third: a
- * name is what to read off a request, a template is what to do with what
- * was read. A door that took the pieces separately would make every caller
- * re-assemble the family, and one of them would forget a rung.
+ * Nothing else in this package touches `process.env`. {@link ./identity.ts}
+ * is a parse over header names it is HANDED and {@link ./picture.ts} is a
+ * ladder over a template it is HANDED — both pure functions of their
+ * arguments, which is what lets a test state a deployment instead of
+ * arranging one, and what stops "how this olai is wired" from being
+ * answered in two places.
+ *
+ * One family, `OLAI_IDENTITY_*`, and it is deliberately not one variable
+ * per proxy: a login, an email claim, a display name, a picture, and the
+ * avatar template that pictures a login when no proxy sent a picture. The
+ * defaults are `tailscale serve`'s own headers, because that is the
+ * deployment olai documents first.
+ *
+ * Read on demand rather than at import, so what a process was STARTED with
+ * is what it serves.
  */
 
 import {
   DEFAULT_IDENTITY_HEADERS,
-  identityHeaders,
+  DEFAULT_LOGIN_HEADER,
+  DEFAULT_NAME_HEADER,
+  DEFAULT_PICTURE_HEADER,
   type IdentityHeaders,
 } from "./identity.ts"
-import { avatarTemplate } from "./picture.ts"
+
+/** The login header's name. Unset is {@link DEFAULT_LOGIN_HEADER}. */
+export const LOGIN_ENV = "OLAI_IDENTITY_LOGIN_HEADER"
+/** The email header's name. Unset is the login header — on the tailnets
+ *  where the login IS an address, one name covers both; empty is no email
+ *  claim, and then nothing is ever hashed into a gravatar. */
+export const EMAIL_ENV = "OLAI_IDENTITY_EMAIL_HEADER"
+/** The display-name header's name. Unset is {@link DEFAULT_NAME_HEADER};
+ *  empty is no name, and then the login is what the chip says. */
+export const NAME_ENV = "OLAI_IDENTITY_NAME_HEADER"
+/** The picture header's name. Unset is {@link DEFAULT_PICTURE_HEADER};
+ *  empty is no picture header, and the ladder starts a rung lower. */
+export const PICTURE_ENV = "OLAI_IDENTITY_PICTURE_HEADER"
+/** The avatar URL template — one URL with `{login}` in it
+ *  ({@link LOGIN_PLACEHOLDER}). Unset or blank is no template. */
+export const AVATAR_ENV = "OLAI_IDENTITY_AVATAR_TEMPLATE"
 
 export interface IdentityConfig {
   /** Which headers this server trusts for who is looking. */
@@ -32,9 +58,36 @@ export const DEFAULT_IDENTITY_CONFIG: IdentityConfig = {
   avatarTemplate: null,
 }
 
-/** What this process was started with. Read on demand, not at import, so
- *  what a process was started with is what it serves. */
+/** One optional name out of the environment: unset is `fallback`, empty
+ *  (or blank) is off, anything else is what it says. The one shape every
+ *  variable in the family has. */
+const named = (variable: string, fallback: string | null): string | null => {
+  const asked = process.env[variable]
+  if (asked === undefined) return fallback
+  const name = asked.trim()
+  return name === "" ? null : name
+}
+
+/**
+ * The header names this process trusts, or the Tailscale defaults.
+ *
+ * The login is the only one that cannot be turned off — it is what makes
+ * somebody present — so it falls back to its default rather than to
+ * `null`, and the email claim falls back to whatever the login header is
+ * called.
+ */
+export const identityHeaders = (): IdentityHeaders => {
+  const login = process.env[LOGIN_ENV]?.trim() || DEFAULT_LOGIN_HEADER
+  return {
+    login,
+    email: named(EMAIL_ENV, login),
+    name: named(NAME_ENV, DEFAULT_NAME_HEADER),
+    picture: named(PICTURE_ENV, DEFAULT_PICTURE_HEADER),
+  }
+}
+
+/** What this process was started with. */
 export const identityConfig = (): IdentityConfig => ({
   headers: identityHeaders(),
-  avatarTemplate: avatarTemplate(),
+  avatarTemplate: named(AVATAR_ENV, null),
 })
