+++ packages/identity/src/identity.test.ts
@@ -1,30 +1,18 @@
 /**
- * Who this request is: the header names, the parse.
+ * Who this request is: the parse.
  *
- * Two layers, and they are not the same test. {@link identityHeaders} is
- * the config — Tailscale by default, Caddy / Authelia by renaming the
- * family. {@link identityOf} is the seam — present, absent, blank,
- * doubled, a login without any claim beside it. WHICH picture the person
- * then wears is {@link ./picture.test.ts}; the hash is
- * {@link ./gravatar.test.ts}. HTTP is `@olai/server`'s, because a door is
- * not a person.
+ * A function of VALUES — the header names it is handed, and the headers a
+ * request arrived with. Present, absent, blank, doubled, a login with no
+ * claim beside it. Which names a deployment actually uses is the
+ * environment's, and so is its test ({@link ./config.test.ts}); WHICH
+ * picture the person then wears is {@link ./picture.test.ts}, and the hash
+ * is {@link ./gravatar.test.ts}. HTTP is `@olai/server`'s, because a door
+ * is not a person.
  */
 
-import { afterEach, expect, test } from "bun:test"
+import { expect, test } from "bun:test"
 
-import {
-  DEFAULT_IDENTITY_HEADERS,
-  DEFAULT_LOGIN_HEADER,
-  DEFAULT_NAME_HEADER,
-  DEFAULT_PICTURE_HEADER,
-  EMAIL_ENV,
-  identityHeaders,
-  identityOf,
-  LOGIN_ENV,
-  NAME_ENV,
-  PICTURE_ENV,
-  type IdentityHeaders,
-} from "./identity.ts"
+import { identityOf, type IdentityHeaders } from "./identity.ts"
 
 const ADA = "ada@example.com"
 
@@ -35,76 +23,6 @@ const AUTHELIA: IdentityHeaders = {
   name: "Remote-Name",
   picture: null,
 }
-
-const VARIABLES = [LOGIN_ENV, EMAIL_ENV, NAME_ENV, PICTURE_ENV] as const
-const was = new Map(VARIABLES.map((key) => [key, process.env[key]]))
-afterEach(() => {
-  for (const key of VARIABLES) {
-    const before = was.get(key)
-    if (before === undefined) delete process.env[key]
-    else process.env[key] = before
-  }
-})
-
-const nothingSet = (): void => {
-  for (const key of VARIABLES) delete process.env[key]
-}
-
-test("unset is tailscale serve: its four headers, login doubling as email", () => {
-  nothingSet()
-  expect(identityHeaders()).toEqual(DEFAULT_IDENTITY_HEADERS)
-  expect(DEFAULT_IDENTITY_HEADERS).toEqual({
-    login: DEFAULT_LOGIN_HEADER,
-    email: DEFAULT_LOGIN_HEADER,
-    name: DEFAULT_NAME_HEADER,
-    picture: DEFAULT_PICTURE_HEADER,
-  })
-})
-
-test("OLAI_IDENTITY_LOGIN_HEADER is the name, and the email follows it", () => {
-  nothingSet()
-  process.env[LOGIN_ENV] = "Remote-User"
-  expect(identityHeaders()).toEqual({
-    login: "Remote-User",
-    email: "Remote-User",
-    name: DEFAULT_NAME_HEADER,
-    picture: DEFAULT_PICTURE_HEADER,
-  })
-})
-
-test("an email header is a second name; empty is no email claim", () => {
-  nothingSet()
-  process.env[LOGIN_ENV] = "Remote-User"
-  process.env[EMAIL_ENV] = "Remote-Email"
-  expect(identityHeaders().email).toBe("Remote-Email")
-  process.env[EMAIL_ENV] = ""
-  expect(identityHeaders().email).toBeNull()
-  process.env[EMAIL_ENV] = "  "
-  expect(identityHeaders().email).toBeNull()
-})
-
-test("the name and picture headers are configurable the same way", () => {
-  nothingSet()
-  process.env[NAME_ENV] = "X-Auth-Request-Preferred-Username"
-  process.env[PICTURE_ENV] = "X-Pomerium-Claim-Picture"
-  expect(identityHeaders().name).toBe("X-Auth-Request-Preferred-Username")
-  expect(identityHeaders().picture).toBe("X-Pomerium-Claim-Picture")
-  process.env[NAME_ENV] = ""
-  process.env[PICTURE_ENV] = "  "
-  expect(identityHeaders().name).toBeNull()
-  expect(identityHeaders().picture).toBeNull()
-})
-
-test("the environment is read per call, not once at import", () => {
-  nothingSet()
-  process.env[LOGIN_ENV] = "X-Auth-Request-User"
-  process.env[EMAIL_ENV] = "X-Auth-Request-Email"
-  expect(identityHeaders().login).toBe("X-Auth-Request-User")
-  process.env[LOGIN_ENV] = "X-Token-User-Nick"
-  process.env[EMAIL_ENV] = "X-Token-User-Email"
-  expect(identityHeaders().login).toBe("X-Token-User-Nick")
-  expect(identityHeaders().email).toBe("X-Token-User-Email")
-})
 
 test("a present login is that login, trimmed, with the email claim", () => {
   expect(identityOf({ "tailscale-user-login": "  ada@example.com  " })).toEqual({
